+++ packages/terra-core-docs/src/terra-dev-site/doc/arrange/common/examplesetup.jsx
@@ -4,7 +4,7 @@ import styles from './examplesetup.scss';
 
 const cx = classNames.bind(styles);
 
-const ipsum = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.';
+const ipsum = 'This example provides content containers with a fit (start and/or end region) and fill (middle region). Arrange Props have one required prop which is fill to the content to display in the body of the fill.';
 const simpleText = <div>{ipsum}</div>;
 const textWithBlueBorder = <div className={cx('outlined-placeholder')}>{ipsum}</div>;