+++ packages/terra-core-docs/src/terra-dev-site/doc/form-radio/example/radio/InlineRadiosExample.jsx
@@ -3,9 +3,9 @@ import Radio from 'terra-form-radio';
 
 const inlineRadiosExample = () => (
   <div>
-    <Radio id="first-inline" labelText="First Radio" name="inline-example" isInline />
-    <Radio id="second-inline" labelText="Second Radio" name="inline-example" isInline />
-    <Radio id="third-inline" labelText="Third Radio" name="inline-example" isInline />
+    <Radio id="inDesign" labelText="InDesign" name="inline-example" isInline />
+    <Radio id="photoshop" labelText="Photoshop" name="inline-example" isInline />
+    <Radio id="illustrator" labelText="Illustrator" name="inline-example" isInline />
   </div>
 );